requirements.txt
streamlit==1.31.0
gitpython==3.1.41
sentence-transformers==2.2.2
faiss-cpu==1.8.0
groq>=0.9.0
python-dotenv==1.0.1
scikit-learn>=1.0.0
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
tiktoken>=0.5.0
